+++ src/app/admin-module/components/recipe-list/recipe-list.component.ts
@@ -1,9 +1,9 @@
 import { Component, ViewEncapsulation } from '@angular/core';
 import { PageEvent } from '@angular/material/paginator';
 
-import { LanguageService } from '../../shared/services/language-service/language.service';
-import locale from '../admin.locale.json';
-import { DataCollectionService } from '../services/data-collection.service';
+import { LanguageService } from '../../../shared/services/language-service/language.service';
+import locale from '../../admin.locale.json';
+import { DataCollectionService } from '../../services/data-collection.service';
 
 @Component({
   selector: 'app-recipe-list',